backend/requirements.txt
boto3==1.10.39
botocore==1.13.39
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-allauth==0.40.0
django-cors-headers==3.2.0
django-heroku==0.3.1
django-storages==1.8
djangorestframework==3.10.3
docutils==0.15.2
gunicorn==19.0.0
html5lib==1.0.1
idna==2.8
jmespath==0.9.4
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.4
PyPDF2==1.26.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.3
reportlab==3.5.32
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
webencodings==0.5.1
whitenoise==4.1.4
xhtml2pdf==0.2.3
